pyproject.toml
[tool.poetry]
name = "Emoji Mate"
version = "2.2.0"
description = "Quickly search for emojis"
documentation = "https://github.com/fedecalendino/alfred-emoji-mate/blob/main/README.md"
homepage = "https://github.com/fedecalendino/alfred-emoji-mate"
license = "MIT"

authors = [
    "Fede Calendino <fede@calendino.com>"
]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
alfred-pyflow = "0.4.7"
fuzzywuzzy = "^0.18.0"
emoji = "^2.12.1"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
coverage = "^7.2.5"
ddt = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
